+++ src/app/app.module.ts
@@ -1,6 +1,8 @@
-import { NgModule } from '@angular/core';
+import { DEFAULT_CURRENCY_CODE, LOCALE_ID, NgModule } from '@angular/core';
 import { BrowserModule } from '@angular/platform-browser';
 import { provideAnimationsAsync } from '@angular/platform-browser/animations/async';
+import { registerLocaleData } from '@angular/common';
+import localePl from '@angular/common/locales/pl';
 
 import { AppRoutingModule } from './app-routing.module';
 import { AngularMaterialModule } from './shared/angular-material.module';
@@ -12,6 +14,8 @@ import { PeriodSelectorComponent } from './components/period-selector/period-sel
 import { SummaryComponent } from './components/summary/summary.component';
 import { TableComponent } from './components/table/table.component';
 
+registerLocaleData(localePl);
+
 @NgModule({
   declarations: [
     AppComponent,
@@ -22,7 +26,17 @@ import { TableComponent } from './components/table/table.component';
     MainPageComponent,
   ],
   imports: [BrowserModule, AppRoutingModule, AngularMaterialModule],
-  providers: [provideAnimationsAsync()],
+  providers: [
+    provideAnimationsAsync(),
+    {
+      provide: LOCALE_ID,
+      useValue: 'pl',
+    },
+    {
+      provide: DEFAULT_CURRENCY_CODE,
+      useValue: 'PLN',
+    },
+  ],
   bootstrap: [AppComponent],
 })
 export class AppModule {}